+++ Biomass.hpp
@@ -18,12 +18,12 @@ protected:
 
 public:
     int max_age_of_cell = simulation_config::biomass::default_max_age;
-    float level_of_resistance = simulation_config::biomass::default_resistance;
+    double level_of_resistance = simulation_config::biomass::default_resistance;
 
     virtual ~abstract_Biomass() = default;
 
     int get_age() const;
-    float get_level_of_resistance() const;
+    double get_level_of_resistance() const;
     double get_biomass() const;
 
     // Теперь must_he_die принимает антибиотик
@@ -36,7 +36,7 @@ public:
     virtual bool reproduction(Field& current_field, int x, int y);
 
     virtual void consume_and_decay(Food& food);
-    virtual float is_active_for_monod() const = 0;
+    virtual double is_active_for_monod() const = 0;
     virtual double get_maintenance_rate() const = 0;
 
     virtual void step_after_death();
@@ -48,31 +48,31 @@ private:
     int steps_active = 0;
 public:
     active_Biomass() = default;
-    active_Biomass(float resistance, int max_age);
+    active_Biomass(double resistance, int max_age);
 
     bool is_alive() const override;
     bool reproduction(Field& current_field, int x, int y) override;
     void consume_and_decay(Food& food) override;
-    float is_active_for_monod() const override { return 1.0f; }
+    double is_active_for_monod() const override { return 1.0; }
     double get_maintenance_rate() const override { return simulation_config::monod::m_act; }
 };
 
 class nonactive_Biomass : public abstract_Biomass {
 private:
-    float resistance_multiplier = simulation_config::biomass::nonactive_resistance_multiplier;
-    float max_life_multiplier = simulation_config::biomass::nonactive_max_life_multiplier;
+    double resistance_multiplier = simulation_config::biomass::nonactive_resistance_multiplier;
+    double max_life_multiplier = simulation_config::biomass::nonactive_max_life_multiplier;
     int steps_until_wakeup = 0;
 public:
     nonactive_Biomass() = default;
     nonactive_Biomass(
-        float resistance,
+        double resistance,
         int max_age
     );
 
     bool is_alive() const override { return true; }
     bool reproduction(Field& current_field, int x, int y) override;
     
-    float is_active_for_monod() const override { return 0.0f; }
+    double is_active_for_monod() const override { return 0.0; }
     double get_maintenance_rate() const override { return simulation_config::monod::m_inactiv; }
 
     // Применяет коэффициенты дормантности (устойчивость/возраст/расход пищи)
@@ -82,7 +82,7 @@ public:
 
     // Обратные величины для восстановления "обычных" параметров клетки
     // при переходе nonactive -> active.
-    float baseline_resistance() const;
+    double baseline_resistance() const;
     int baseline_max_age() const;
 };
 
@@ -92,7 +92,7 @@ private:
 public:
     int count_of_steps_from_death = 0;
     
-    float is_active_for_monod() const override { return 0.0f; }
+    double is_active_for_monod() const override { return 0.0; }
     double get_maintenance_rate() const override { return 0.0; }
     
     void step_after_death() override;
